Do manyROC analysis with cross-validation for hyperSpec object

[!!!] // No description yet //

sp_manyroc_with_cv(Var, Spectra, k_folds = 5, times = 10,
  seeds = 2222222, kind = "L'Ecuyer-CMRG", n_min = k_folds)

Arguments

Var
(character(1)\ factor)
Either the name of variable in Spectra which contains the grouping variable or a factor vector (or convertible to factor) with values for grouping.

Spectra
hyperSpec object, that contains colum ID for spectra blocking, columns with grouping variables as well as spectroscopic information.

k_folds
(positive integer)
The number of folds in k-fold cross-validation.

times
(positive integer)
The number of repetitions in repeated k-fold cross-validation.

seeds
(NULL | integer)
Either a vector of integers of length times to set seed for each repetition of k-fold cross-validation. For more about seeds see set.seed().

Each seed will be passed to cvo_create_folds.()

kind
character or NULL. If kind is a character string, set R's RNG to the kind desired. Use "default" to return to the R default. See ‘Details’ for the interpretation of NULL.

n_min
integer(1)
minimum acceptable number of unique samples per group. Must be at least k_folds or bigger.

Details

Function sp_manyroc_with_cv will be renamed in the future.

See also

Examples

library(manyROC)

fluorescence$ID  <- 1:nrow(fluorescence)
sp_manyroc_with_cv("gr", fluorescence[,,500~502], k_folds = 3, times = 2)

#> Error in distinct(DF, .keep_all = TRUE): could not find function "distinct"