Do manyROC analysis with cross-validation for hyperSpec object

[!!!] // No description yet //

sp_manyroc_with_cv(Var, Spectra, k_folds = 5, times = 10, seeds = 2222222,
  kind = "L'Ecuyer-CMRG", n_min = k_folds)

Arguments

Var
(character(1)\ factor) Either the name of variable in Spectra which contains the grouping variable or a factor vector (or convertible to factor) with values for grouping.

Spectra
hyperSpec object, that contains colum ID for spectra blocking, columns with grouping variables as well as spectroscopic information.

k_folds
(positive integer) The number of folds in k-fold cross-validation.

times
(positive integer) The number of repetitions in repeated k-fold cross-validation.

seeds
(NULL | integer) Either a vector of integers of length times to set seed for each repetition of k-fold cross-validation. For more about seeds see set.seed().

Each seed will be passed to cvo_create_folds.()

kind
character or NULL. If kind is a character string, set R's RNG to the kind desired. Use "default" to return to the R default. See ‘Details’ for the interpretation of NULL.

n_min
integer(1) minimum acceptable number of unique samples per group. Must be at least k_folds or bigger.

Details

Function sp_manyroc_with_cv will be renamed in the future.

See also

sp_manyroc_with_cv_by_variable

Examples

library(manyROC)

fluorescence$ID  <- 1:nrow(fluorescence)
sp_manyroc_with_cv("gr", fluorescence[,,500~502], k_folds = 3, times = 2)

#> $variable
#> [1] "gr"
#> 
#> $n_included
#> [1] 150
#> 
#> $ind_included_rows
#> Classes 'as_str', 'logical'  logi [1:150] TRUE TRUE TRUE TRUE TRUE TRUE ...
#> 
#> $cvo
#> --- A cvo object: ----------------------------------------------------
#>  indices stratified blocked         cv_type k repetitions sample_size
#>    Train       TRUE   FALSE Repeated k-fold 3           2         150
#> 
#> Random number generator: L'Ecuyer-CMRG 
#> Seeds:  2222222, 24997039
#> ----------------------------------------------------------------------
#> 
#> $results
#>          set       fold compared_groups feature median_neg cutoff median_pos
#> 1   training Rep1_Fold1         A vs. B     500        336    369        333
#> 2   training Rep1_Fold1         A vs. C     500        336    343        343
#> 3   training Rep1_Fold1         B vs. C     500        333    343        343
#> 4   training Rep1_Fold2         A vs. B     500        339    348        333
#> ...      ...        ...             ...     ...        ...    ...        ...
#> 105     test Rep2_Fold2         B vs. C     502         NA    340         NA
#> 106     test Rep2_Fold3         A vs. B     502         NA    350         NA
#> 107     test Rep2_Fold3         A vs. C     502         NA    319         NA
#> 108     test Rep2_Fold3         B vs. C     502         NA    339         NA
#>                                                                             
#>      tp  fn  fp  tn sens spec  ppv  npv  bac youden kappa  auc
#> 1    33   3  26   8 0.92 0.24 0.56 0.73 0.58   0.15  0.15 0.54
#> 2    15  14  11  23 0.52 0.68 0.58 0.62 0.60   0.19  0.20 0.54
#> 3    15  14  10  26 0.52 0.72 0.60 0.65 0.62   0.24  0.24 0.58
#> 4    31   6  21  14 0.84 0.40 0.60 0.70 0.62   0.24  0.24 0.60
#> ... ... ... ... ...  ...  ...  ...  ...  ...    ...   ...  ...
#> 105   8   6   5  14 0.57 0.74 0.62 0.70 0.65   0.31  0.31   NA
#> 106  13   5  13   4 0.72 0.24 0.50 0.44 0.48  -0.04 -0.04   NA
#> 107   3  11   8   9 0.21 0.53 0.27 0.45 0.37  -0.26 -0.26   NA
#> 108   8   6   5  13 0.57 0.72 0.62 0.68 0.65   0.29  0.30   NA
#>                                                               
#>