Do manyROC analysis with cross-validation for hyperSpec object for each variable

[!!!] // No description yet //

For reproducible results in parallel computing, set seed with parallel::clusterSetRNGStream(iseed = x)(where x is your seed) (and not with set.seed) as package parallelMap is used for paralellization.

# S3 method for hide_it
print(x, ...)

sp_manyroc_with_cv_by_variable(
  Spectra,
  variables_to_analyze,
  k_folds = 3,
  times = 10,
  seeds = NULL,
  kind = NULL
)

Arguments

x
an object used to select a method.

...
further arguments passed to or from other methods.

Spectra
hyperSpec object, that contains colum ID for spectra blocking, columns with grouping variables as well as spectroscopic information.

variables_to_analyze
(character)
A character vector with names of grouping variables to use in analysis.

k_folds
(positive integer)
The number of folds in k-fold cross-validation.

times
(positive integer)
The number of repetitions in repeated k-fold cross-validation.

seeds
(NULL | integer)
Either a vector of integers of length times to set seed for each repetition of k-fold cross-validation. For more about seeds see set.seed().

Each seed will be passed to cvo_create_folds.()

kind
character or NULL. If kind is a character string, set R's RNG to the kind desired. Use "default" to return to the R default. See ‘Details’ for the interpretation of NULL.

Value

A list with results
[!!!] Description needs more specification.

Details

[!!!] The seeds nust be at least of length 6 and meet other requirements for "L'Ecuyer-CMRG" random number generator.

[!!!] At the moment Seeding section needs revision if it is necessary to use "L'Ecuyer-CMRG" (pseudo)random number generator as set.seed() does not work with parallelMap package. Instead parallel::clusterSetRNGStream(iseed = x) (where x is your seed) should be used with parallelMap to get reproducible results.

See also

Examples

library(manyROC)

fluorescence$ID  <- 1:nrow(fluorescence)
sp_manyroc_with_cv_by_variable(
        fluorescence[,,500~501],
         c("gr", "class"),
         k_folds = 3,
         times = 2)

#> Exporting objects to package env on master for mode: local

#> Error in class(x) <- union(c(new_class, class(x)), class(x)): attempt to set an attribute on NULL