Do manyROC analysis with cross-validation for hyperSpec object for each variable

[!!!] // No description yet //

sp_manyroc_with_cv_by_variable(Spectra, variables_to_analyze, k_folds = 3,
  times = 10, seeds = 2222222)

Arguments

Spectra
hyperSpec object, that contains colum ID for spectra blocking, columns with grouping variables as well as spectroscopic information.

variables_to_analyze
(character) A character vector with names of grouping variables to use in analysis.

k_folds
(positive integer) The number of folds in k-fold cross-validation.

times
(positive integer) The number of repetitions in repeated k-fold cross-validation.

seeds
(NULL | integer) Either a vector of integers of length times to set seed for each repetition of k-fold cross-validation. The seeds nust be at least of length 6 and meet other requirements for "L'Ecuyer-CMRG" random number generator.

Value

A list with results [!!!] Description needs more specification.

See also

sp_manyroc_with_cv

Examples

library(manyROC)

fluorescence$ID  <- 1:nrow(fluorescence)
sp_manyroc_with_cv_by_variable(fluorescence[,,500~501], c("gr", "class"),
                          k_folds = 3, times = 2)

#> $n_included
#>    gr class
#> 1 150   150
#> 
#> $ind_included_rows
#>       gr class
#> 1   TRUE  TRUE
#> 2   TRUE  TRUE
#> 3   TRUE  TRUE
#> 4   TRUE  TRUE
#> ...  ...   ...
#> 147 TRUE  TRUE
#> 148 TRUE  TRUE
#> 149 TRUE  TRUE
#> 150 TRUE  TRUE
#>               
#> 
#> $cvo
#> *** First non-empty element: ***
#> --- A cvo object: ----------------------------------------------------
#>  indices stratified blocked         cv_type k repetitions sample_size
#>    Train       TRUE   FALSE Repeated k-fold 3           2         150
#> 
#> Random number generator: L'Ecuyer-CMRG 
#> Seeds:  2222222, 24997039
#> ----------------------------------------------------------------------
#> 
#> *** Other elements are not shown ***
#> 
#> 
#> $results
#>     grouping      set       fold compared_groups feature median_neg cutoff
#> 1         gr training Rep1_Fold1         A vs. B     500        336    369
#> 2         gr training Rep1_Fold1         A vs. C     500        336    343
#> 3         gr training Rep1_Fold1         B vs. C     500        333    343
#> 4         gr training Rep1_Fold2         A vs. B     500        339    348
#> ...      ...      ...        ...             ...     ...        ...    ...
#> 213    class     test Rep2_Fold3        K vs. S1     501         NA    327
#> 214    class     test Rep2_Fold3         l vs. N     501         NA    348
#> 215    class     test Rep2_Fold3        l vs. S1     501         NA    346
#> 216    class     test Rep2_Fold3        N vs. S1     501         NA    328
#>                                                                           
#>     median_pos  tp  fn  fp  tn sens spec  ppv  npv  bac youden kappa  auc
#> 1          333  33   3  26   8 0.92 0.24 0.56 0.73 0.58   0.15  0.15 0.54
#> 2          343  15  14  11  23 0.52 0.68 0.58 0.62 0.60   0.19  0.20 0.54
#> 3          343  15  14  10  26 0.52 0.72 0.60 0.65 0.62   0.24  0.24 0.58
#> 4          333  31   6  21  14 0.84 0.40 0.60 0.70 0.62   0.24  0.24 0.60
#> ...        ... ... ... ... ...  ...  ...  ...  ...  ...    ...   ...  ...
#> 213         NA   8   9   9   8 0.47 0.47 0.47 0.47 0.47  -0.06 -0.06   NA
#> 214         NA   4   3   7   3 0.57 0.30 0.36 0.50 0.44  -0.13 -0.12   NA
#> 215         NA  11   6   7   3 0.65 0.30 0.61 0.33 0.47  -0.05 -0.05   NA
#> 216         NA   8   9   2   5 0.47 0.71 0.80 0.36 0.59   0.18  0.14   NA
#>                                                                          
#> 
#> $variables_included
#> [1] "gr"    "class"
#> 
#> $variables_errored
#> character(0)
#> 
#> $error_messages
#> *** First non-empty element: ***
#> NULL
#> 
#> *** Other elements are not shown ***
#> 
#>